Fix next/headers error by separating server and client components
