+++ web/src/app/dashboard/layout.tsx
@@ -1,11 +1,9 @@
-"use client";
-
 import { DashboardNav } from "@/components/dashboard/DashboardNav";
 import { Sidebar } from "@/components/dashboard/Sidebar";
 import { createServerClient } from "@supabase/ssr";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
-import { useState, useEffect } from "react";
+import { DashboardLayoutClient } from "@/components/dashboard/DashboardLayoutClient";
 
 export default async function DashboardLayout({
   children,
@@ -35,56 +33,5 @@ export default async function DashboardLayout({
     redirect("/auth/login");
   }
 
-  return <DashboardLayoutClient user={user} children={children} />;
-}
-
-// Client component to handle state
-function DashboardLayoutClient({
-  user,
-  children,
-}: {
-  user: any;
-  children: React.ReactNode;
-}) {
-  const [isSidebarOpen, setIsSidebarOpen] = useState(false);
-
-  const toggleSidebar = () => {
-    setIsSidebarOpen(!isSidebarOpen);
-  };
-
-  // Close sidebar when window is resized to desktop size
-  useEffect(() => {
-    const handleResize = () => {
-      if (window.innerWidth >= 1024) {
-        setIsSidebarOpen(false);
-      }
-    };
-
-    window.addEventListener('resize', handleResize);
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
-
-  return (
-    <div className="min-h-screen bg-background flex">
-      {/* Sidebar */}
-      <Sidebar isCollapsed={!isSidebarOpen} onToggle={toggleSidebar} />
-
-      {/* Mobile Overlay */}
-      {isSidebarOpen && (
-        <div 
-          className="fixed inset-0 bg-black/50 z-40 lg:hidden" 
-          onClick={toggleSidebar}
-          aria-hidden="true"
-        />
-      )}
-
-      {/* Main Content */}
-      <div className="flex-1 flex flex-col min-h-screen overflow-hidden w-full">
-        <DashboardNav user={user} toggleSidebar={toggleSidebar} />
-        <main className="flex-1 p-4 sm:p-6 bg-background overflow-auto">
-          <div className="max-w-7xl mx-auto">{children}</div>
-        </main>
-      </div>
-    </div>
-  );
+  return <DashboardLayoutClient user={user}>{children}</DashboardLayoutClient>;
 }
