+++ web/src/components/dashboard/DashboardLayoutClient.tsx
@@ -1,0 +1,55 @@
+"use client";
+
+import { useState, useEffect } from "react";
+import { DashboardNav } from "@/components/dashboard/DashboardNav";
+import { Sidebar } from "@/components/dashboard/Sidebar";
+
+export function DashboardLayoutClient({
+  user,
+  children,
+}: {
+  user: any;
+  children: React.ReactNode;
+}) {
+  const [isSidebarOpen, setIsSidebarOpen] = useState(false);
+
+  const toggleSidebar = () => {
+    setIsSidebarOpen(!isSidebarOpen);
+  };
+
+  // Close sidebar when window is resized to desktop size
+  useEffect(() => {
+    const handleResize = () => {
+      if (window.innerWidth >= 1024) {
+        setIsSidebarOpen(false);
+      }
+    };
+
+    window.addEventListener('resize', handleResize);
+    return () => window.removeEventListener('resize', handleResize);
+  }, []);
+
+  return (
+    <div className="min-h-screen bg-background flex">
+      {/* Sidebar */}
+      <Sidebar isCollapsed={!isSidebarOpen} onToggle={toggleSidebar} />
+
+      {/* Mobile Overlay */}
+      {isSidebarOpen && (
+        <div 
+          className="fixed inset-0 bg-black/50 z-40 lg:hidden" 
+          onClick={toggleSidebar}
+          aria-hidden="true"
+        />
+      )}
+
+      {/* Main Content */}
+      <div className="flex-1 flex flex-col min-h-screen overflow-hidden w-full">
+        <DashboardNav user={user} toggleSidebar={toggleSidebar} />
+        <main className="flex-1 p-4 sm:p-6 bg-background overflow-auto">
+          <div className="max-w-7xl mx-auto">{children}</div>
+        </main>
+      </div>
+    </div>
+  );
+}
